Move auth to directory routers and schemas
