+++ src/db/database.py
@@ -20,11 +20,3 @@ AsyncSessionLocal = sessionmaker(
 )
 
 Base = declarative_base()
-
-# async def init_db():
-#     async with engine.begin() as conn:
-#         await conn.run_sync(Base.metadata.create_all)
-
-# async def get_db() -> AsyncGenerator[AsyncSession, None]:
-#     async with AsyncSessionLocal() as session:
-#         yield session
