+++ src/main.py
@@ -1,10 +1,10 @@
 
 from fastapi import FastAPI
 
-from src import auth
+from src.routers.auth_routers import router as auth_router
 from src.routers.main_routes import router as main_router
 
 
 app = FastAPI()
-app.include_router(auth.router)
+app.include_router(auth_router)
 app.include_router(main_router)
