+++ src/schemas/auth_schemas.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+class CreateUserRequest(BaseModel):
+    username: str
+    password: str
+    email: str
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
